Take the wall clock out of the orphan recovery tests
The recovery tests slept for a real 450ms per settle() — "plus slack for a
loaded CI machine" — which is exactly what AGENTS.md forbids: a padded wait
threshold is a guess that flakes when CI beats it and then invites a bigger
guess. Several tests call settle() more than once, so the file was adding
seconds of nondeterministic wall clock on its own.

The adjacent loop and offset tests already had the right answer in the form of a
local FakeClock. Extracted it to test/fake-clock.ts and pointed all three files
at it, removing two duplicate copies along the way.

Still load-bearing: breaking the recovery branch fails the test in 1.04ms rather
than after a sleep. The editor suite drops from 4.12s to 1.86s.
